Move NMS and NMM related methods from utils into overlap_handling.py
